fix(planner): declare integer arithmetic integer wherever it is nested
`MIN(CASE WHEN g = 0 THEN v ELSE v + 100 END)` over a bigint column answered
float64 where PostgreSQL 17 answers bigint.

The integer rule has existed since #369, but only in `inferProjectionDeclType`,
and there it fires only for the OUTERMOST node of a projection. Every nested
site — a CASE branch, a COALESCE/GREATEST/LEAST argument, an aggregate's input
— goes through `nodeDeclaredType`, whose BinaryOp arm returned
`Decl(TypeFloat64), Decided` for anything not DECIMAL and not temporal. One
expression, two declarations, decided by where it sat: `SELECT v + 100`
declared INT64 and `CASE … ELSE v + 100 END` declared FLOAT64.

It read as correct until the fold stopped being order-dependent.
`CommonDeclType` used to answer from the first non-DECIMAL decider, so
{INT64, FLOAT64} returned INT64 by ARGUMENT ORDER — writing the arms the
other way round already answered float8 before the v0.18.6 stack. e61f0a4e
replaced that with PostgreSQL's select_common_type ladder, which folds
{INT64, FLOAT64} to FLOAT64 correctly and so exposed the FLOAT64 this arm had
been contributing all along. The ladder is right; its INPUT was wrong.

The arm now consults `intArithAllInt` — the same predicate the projection
uses, and a strict subset of expr.BinOpNumeric's runtime integer mode, so the
declaration cannot promise an integer the kernel will not produce. It rides
`expr.IntArithOn()`: with WADJET_INT_ARITH=0 the kernel takes its float
delegate and declaring INT64 would be the corrupting direction.
`intArithAllInt` gains one column authority for the new caller — a nil
strictInt reads `decls`, which is what `colRefDeclaredType` already trusts to
type a bare column at these very sites. The existing caller is byte-identical.

PostgreSQL 17 matrix behind this, read off a live server (format_type over a
view, so typmods are carried), eight ARM KINDS x five widths — bare column,
arithmetic, CAST, integer/numeric/quoted literal, nested CASE, function call.
All 40 rows are now in `coordinator.nfCorpus`, run on BOTH paths, comparing
the value AND the Go box. 32 agree outright. The 8 that do not ALL reproduce
with this change reverted, which is how they were classified rather than
assumed, and each is now pinned with PostgreSQL's answer recorded beside it:

  * int32 arithmetic declares INT64 (PostgreSQL: integer) — integer
    arithmetic has one width in this engine, the widening LiteralTypes
    already records at the wire;
  * ABS() over an integer or a real declares float8 (PostgreSQL preserves the
    argument's domain) — expr.DefaultRegistry's return declarations, a
    different site this does not reach, and a wrong NUMBER for real:
    0.1 comes back 0.10000000149011612;
  * a bare numeric-literal arm beside an integer column resolves to the
    COLUMN's type, so 100.5 stores as 100 where PostgreSQL folds the pair to
    numeric;
  * float -> integer CAST rounds the half away from zero where PostgreSQL's
    rint() rounds to even, so -0.5 casts to -1 rather than 0.

The last three are wrong VALUES, and none had a gate. `nfValuePins` is new
and works the way `nfCarrierRefusals` does: the gate asserts each STILL
diverges, so the day one is fixed its pin fails and PostgreSQL's answer is
already in the corpus to assert instead (ADR-0013).

Three existing expectations recorded the FLOAT64 and are corrected, values
unchanged in all three:

  * physical `{"i64 + i32", FLOAT64}` — its own comment said the declaration
    was "the caller's, not this one's", which was the split this closes;
  * ./test `SELECT 5 - 2 AS x` float64 -> int64 (PostgreSQL: integer 3);
  * ./test `SELECT MIN(i64 + 1)` float64 -> int64 (PostgreSQL: bigint 2).

pg-oracle gains five semantics entries and two wire entries. The semantics
one that can FAIL rather than merely document is
`IntegerArithmeticInACaseArmIsExact`: 2^53+1 has no float64, so an arm
declared float8 answers 9007199254740992 — off by one, in a column a client
reads as an integer. It fails on revert. The wire entries' OIDs are pinned
for the int4/int8 width, so the domain is gated by
`TestNumericFoldTwoPath`'s unpinned box assertion instead, which the wire
comment says.

TestDeclaredFoldAgreesWithTheComparisonFold, TestTPCHStageDumpGolden,
TestTPCHQueries and TestTPCHOptimizationInvariance are unaffected: no TPC-H
expression is all-integer arithmetic — `ps_supplycost * ps_availqty` is
FLOAT64 x INT32 and stays FLOAT64.

Gate: `TestMinMaxOverStringCase/min_over_a_numeric_CASE` in ./test, the CI
job that runs ./test. Fails on revert.

Refs #724

+++ test/date_cast_test.go
@@ -331,7 +331,13 @@ func TestDateCastLeavesNonTemporalArithmeticAlone(t *testing.T) {
 		// A time-of-day literal has no column type to become, so it keeps
 		// its text.
 		{"SELECT TIME '10:00:00' AS x", "10:00:00"},
-		{"SELECT 5 - 2 AS x", float64(3)},
+		// Integer arithmetic, so an integer: `SELECT pg_typeof(5 - 2)` is
+		// `integer` on PostgreSQL 17. The width still differs — this engine
+		// computes integer arithmetic in int64 and declares it INT64, which
+		// is what `SELECT 5 - 2 FROM t` has answered since #369 — but the
+		// DOMAIN is no longer float, and it no longer depends on whether the
+		// expression sits at the top of a projection.
+		{"SELECT 5 - 2 AS x", int64(3)},
 	}
 	for _, c := range cases {
 		rows := dateCastRows(t, ctx, db, c.sql)
